+++ src/enums/stash-global-commands.enum.ts
@@ -2,5 +2,7 @@ export enum StashGlobalCommands {
     refresh = 'scm-local-stashes.refresh',
     create = 'scm-local-stashes.create',
     createIncludeUntracked = 'scm-local-stashes.create-include-untracked',
-    createForRepo = 'scm-local-stashes.repo.create'
+    createForRepo = 'scm-local-stashes.repo.create',
+    createFromStaged = 'scm-local-stashes.scm.stash-staged',
+    createFromUnstaged = 'scm-local-stashes.scm.stash-unstaged'
 }