+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 * Add a Stashes view to create, view, apply, pop and drop stashes
+* Add stash buttons to the Staged Changes and Changes sections of the Source Control view
 
 ## [1.6.2] - 2024-01-07